Working traversal for mathematical ops
Had a breakthrough making walkAddition work, this cascaded to make the other 4 [-, *, / and %] also work. As the only difference between the four methods was the opcode that was pushed to the code store post traversal, it made sense to refactor them into a simpler method. Similar refactoring occurred for their respective shorthand assignment operations.