First draft of data model for MVM.
New system uses a tree of stack frames - each knows its parent, and all we
keep track of are the leaf nodes (themselves in a stack...). This provides
a very high level and comprehensible way of laying out the data I hope -
the stack of leaf nodes is much like the return address stack in an x86
system and tree like structure arises from the necessities presented by
block scoping etc. Next step is to integrate this into MVM.