Initial block level scoping.
Variables defined in inner scopes now override those in outer scopes, like
so:

num i = 6;
{
	num i = 7;
	num j = i * i;
}

The outer scope will contain [6], while the inner scope will house [7, 49]
due to scoping rules. This is not a comprehensive implementation - in
fact, it only works in specifically defined blocks right now. However, it
demonstrates that the data models in both MVM and the code generator work
together and function as intended.