Init and render functions.
They're now properly read, and hook up to the driver just as they should.
render is now defined as follows, due to needing a way to import current
time to a render:

function render(num time)->void{
	#time is the number of milliseconds since init was called.
	#...
}

Just one or two more operands, as well as control flow constructs and then
we're done for good.